{"schema":"decodex/commit/1","summary":"Split run control and state internal modules","authority":"manual"}
